requirements.txt
# pinned to avoid surprises; update when you want
plotly>=5.24,<6.0
numpy>=1.26,<2.0
pandas>=2.2,<3.0
scikit-learn>=1.4,<2.0
pyyaml>=6.0.1,<7.0
